fix(training): move SFT builder to active 2b tier
