+++ packages/training/scripts/build_eliza1_sft_0_8b.py
@@ -1,9 +1,13 @@
 #!/usr/bin/env python3
-"""Entry point for the active Eliza-1 0.8B SFT dataset builder."""
+"""Deprecated compatibility entry point for the retired Eliza-1 0.8B SFT builder.
+
+Use ``scripts/build_eliza1_sft_2b.py``. This wrapper now builds the active
+Gemma 4 E2B / ``eliza-1-2b`` SFT dataset.
+"""
 
 from __future__ import annotations
 
-from build_eliza1_sft_0_6b import main
+from build_eliza1_sft_2b import main
 
 
 if __name__ == "__main__":
